+++ test/mocks/ALCXSource.sol
@@ -7,7 +7,7 @@ import {IALCXSource} from "../../src/interfaces/IALCXSource.sol";
 contract ALCXSource is IALCXSource {
 
     IERC20 public alcx = IERC20(0xdBdb4d16EdA451D0503b854CF79D55697F90c8DF);
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function getStakeTotalDeposited(address _user, uint256) external view returns (uint256) {
         return balances[_user];
